+++ examples/GenFitKpipi.cc
@@ -75,6 +75,7 @@ int main( int argc, char** argv )
 	Int_t iFit(0);
 	Int_t nExpt(1);
 	Int_t firstExpt(0);
+	Bool_t isToy(kTRUE);
 	if ( command == "gen" ) {
 		if ( argc > 2 ) {
 			nExpt = atoi( argv[2] );
@@ -234,7 +235,7 @@ int main( int argc, char** argv )
 	LauParameter * signalAsym = new LauParameter("signalAsym",sigAsym,-1.0,1.0,fixSigAsym);
 
 	fitModel->setNSigEvents(signalEvents, signalAsym);
-	fitModel->setNExpts(nExpt, firstExpt);
+	fitModel->setNExpts(nExpt, firstExpt, isToy);
 
 	// Set up the background configuration
 	const Int_t nBBs(1);