+++ examples/GenFitNoDPMultiDim.cc
@@ -66,6 +66,7 @@ int main( int argc, char** argv )
 	Int_t iFit(0);
 	Int_t nExpt(1);
 	Int_t firstExpt(0);
+	Bool_t isToy(kTRUE);
 	if ( command == "gen" ) {
 		if ( argc > 2 ) {
 			nExpt = atoi( argv[2] );
@@ -129,7 +130,7 @@ int main( int argc, char** argv )
 
 	fitModel->setNSigEvents(nSig);
 	fitModel->setNBkgndEvents(nBkg);
-	fitModel->setNExpts(nExpt, firstExpt);
+	fitModel->setNExpts(nExpt, firstExpt, isToy);
 
 
 	// Signal mES PDF parameter values