+++ src/LauFitObject.cc
@@ -28,23 +28,26 @@ Thomas Latham
 
 #include <iostream>
 
+#include "TDecompChol.h"
 #include "TMatrixD.h"
+#include "TRandom.h"
 #include "TSystem.h"
 #include "TVectorD.h"
 
 #include "LauFitObject.hh"
+#include "LauRandom.hh"
 
 ClassImp(LauFitObject)
 
 LauFitObject::LauFitObject() : TObject(),
-	storeCon_(),
 	twoStageFit_(kFALSE),
 	useAsymmFitErrors_(kFALSE),
 	nParams_(0),
 	nFreeParams_(0),
 	withinAsymErrorCalc_(kFALSE),
+	toyExpts_(kFALSE),
 	firstExpt_(0),
-	nExpt_(0),
+	nExpt_(1),
 	iExpt_(0),
 	evtsPerExpt_(0),
 	fitStatus_({-1,0.0,0.0}),
@@ -53,6 +56,19 @@ LauFitObject::LauFitObject() : TObject(),
 	numberOKFits_(0),
 	numberBadFits_(0)
 {
+}
+
+void LauFitObject::setNExpts(UInt_t nExperiments, UInt_t firstExperiment, Bool_t toyExpts)
+{
+	nExpt_ = nExperiments;
+	firstExpt_ = firstExperiment;
+	toyExpts_ = toyExpts;
+
+	if ( ! toyExpts && ( nExperiments > 1 || firstExperiment > 0 ) ) {
+		std::cerr << "WARNING in LauFitObject::setNExpts : toyExpts is set to kFALSE but the values of nExperiments and firstExperiment indicate otherwise, please check" << std::endl;
+	} else if ( toyExpts && nExperiments == 1 && firstExperiment == 0 ) {
+		std::cerr << "WARNING in LauFitObject::setNExpts : toyExpts is set to kTRUE but the values of nExperiments and firstExperiment perhaps indicate otherwise, please check" << std::endl;
+	}
 }
 
 void LauFitObject::resetFitCounters()
@@ -95,63 +111,65 @@ void LauFitObject::storeFitStatus( const LauAbsFitter::FitStatus& status, const 
 
 void LauFitObject::addConstraint(const TString& formula, const std::vector<TString>& pars, const Double_t mean, const Double_t width)
 {
-	if ( ! this->checkRepetition(pars) ){
-		std::cerr << "ERROR in LauFitObject::addConstraint : Parameter(s) added to multiple constraints!" << std::endl;
-		gSystem->Exit( EXIT_FAILURE );
-	}
-
-	StoreConstraints newCon;
-	newCon.formula_ = formula;
-	newCon.conPars_ = pars;
-	newCon.mean_ = mean;
-	newCon.width_ = width;
-
-	storeCon_.push_back(newCon);
-	
-	std::cout << "INFO in LauFitObject::addConstraint : Added formula to constrain" << std::endl;
+	std::cerr << "WARNING in LauFitObject::addConstraint : This function is deprecated, please switch to addFormulaConstraint!" << std::endl;
+	this->addFormulaConstraint( formula, pars, mean, width );
 }
 
-void LauFitObject::addNDConstraint( const std::vector<TString>& pars, const TVectorD& means, const TMatrixD& covMat)
+void LauFitObject::addFormulaConstraint(const TString& formula, const std::vector<TString>& pars, const Double_t mean, const Double_t width)
 {
-	if ( covMat.GetNcols() != covMat.GetNrows() ){
-		std::cerr << "ERROR in LauFitObject::addNDConstraint : Covariance matrix is not square!" << std::endl;
-		gSystem->Exit( EXIT_FAILURE );
-	}
-
-	if ( ( pars.size() != static_cast<std::size_t>(means.GetNrows()) ) || ( pars.size() != static_cast<std::size_t>(covMat.GetNcols()) ) ){
-		std::cerr << "ERROR in LauFitObject::addNDConstraint : Different number of elements in vectors/covariance matrix!" << std::endl; 
-		gSystem->Exit( EXIT_FAILURE );
-	}
-
 	if ( ! this->checkRepetition(pars) ){
-		std::cerr << "ERROR in LauFitObject::addNDConstraint : Parameter(s) added to multiple constraints!" << std::endl;
-		gSystem->Exit( EXIT_FAILURE );
+		std::cerr << "WARNING in LauFitObject::addFormulaConstraint : Parameter(s) added to multiple constraints!" << std::endl;
 	}
 
-	TMatrixD invCovMat {TMatrixD::kInverted, covMat};
-	// Check invertion was successful
-	if ( invCovMat == covMat ){
-		std::cerr << "ERROR in LauFitObject::addNDConstraint : covariance matrix inversion failed, check your input!" << std::endl; 
-		gSystem->Exit( EXIT_FAILURE );
+	formulaConstraints_.emplace_back( FormulaConstraint{formula, pars, mean, width, nullptr} );
+	
+	std::cout << "INFO in LauFitObject::addFormulaConstraint : Added formula constraint" << std::endl;
+}
+
+void LauFitObject::addMultiDimConstraint( const std::vector<TString>& pars, const TVectorD& means, const TMatrixD& covMat)
+{
+	if ( ! this->checkRepetition(pars) ){
+		std::cerr << "WARNING in LauFitObject::addMultiDimConstraint : Parameter(s) added to multiple constraints!" << std::endl;
 	}
 
-	TVectorD values(means.GetNrows());
-	StoreNDConstraints newCon { pars, means, invCovMat, {}, values };
-	storeNDCon_.emplace_back( newCon );
+	multiDimConstraints_.emplace_back( pars, means, covMat );
 
-	std::cout << "INFO in LauFitObject::addNDConstraint : Added list of parameters to constrain" << std::endl;
+	std::cout << "INFO in LauFitObject::addMultiDimConstraint : Added multi-dimensional constraint" << std::endl;
+}
+
+void LauFitObject::generateConstraintMeans( std::vector<LauAbsRValue*>& conVars )
+{
+	if ( ! this->toyExpts() ) {
+		return;
+	}
+
+	// For reproducibility, set a seed based on the experiment number
+	// First, store the current seed, so that it can be restored afterwards
+	const UInt_t oldSeed { LauRandom::randomFun()->GetSeed() };
+	LauRandom::randomFun()->SetSeed( 827375 + this->iExpt() );
+
+	for ( LauAbsRValue* par : conVars ) {
+		par->generateConstraintMean();
+	}
+
+	for ( auto& constraint : multiDimConstraints_ ) {
+		constraint.generateConstraintMeans();
+	}
+
+	// Restore the old random seed
+	LauRandom::randomFun()->SetSeed( oldSeed );
 }
 
 Bool_t LauFitObject::checkRepetition( const std::vector<TString>& names )
 {
 	Bool_t allOK(kTRUE);
 
-	if ( storeCon_.size()==0 && storeNDCon_.size()==0  ) {
+	if ( formulaConstraints_.size()==0 && multiDimConstraints_.size()==0  ) {
 		return allOK;
 	}
 
 	//Check in formula constraints	
-	for ( auto& constraint : storeCon_ ){
+	for ( auto& constraint : formulaConstraints_ ){
 		for ( auto& parname : constraint.conPars_ ){
 			for ( auto& newname : names ){
 				if ( parname == newname ){
@@ -162,7 +180,7 @@ Bool_t LauFitObject::checkRepetition( const std::vector<TString>& names )
 		}
 	}	 	
 	//Check in ND constraints	
-	for ( auto& constraint : storeNDCon_ ){
+	for ( auto& constraint : multiDimConstraints_ ){
 		for ( auto& parname : constraint.conPars_ ){
 			for ( auto& newname : names ){
 				if ( parname == newname ){
@@ -173,4 +191,61 @@ Bool_t LauFitObject::checkRepetition( const std::vector<TString>& names )
 		}
 	}	 
 	return allOK;	
+}
+
+LauFitObject::MultiDimConstraint::MultiDimConstraint( const std::vector<TString>& parNames, const TVectorD& means, const TMatrixD& covMat ) :
+	conPars_{parNames},
+	trueMeans_{means},
+	means_{means},
+	invCovMat_{covMat.GetNrows(),covMat.GetNcols()},
+	sqrtCovMat_{covMat.GetNrows(),covMat.GetNcols()}
+{
+	if ( covMat.GetNcols() != covMat.GetNrows() ){
+		std::cerr << "ERROR in LauFitObject::MultiDimConstraint : Covariance matrix is not square!" << std::endl;
+		gSystem->Exit( EXIT_FAILURE );
+	}
+
+	if ( ( parNames.size() != static_cast<std::size_t>(means.GetNrows()) ) || ( parNames.size() != static_cast<std::size_t>(covMat.GetNrows()) ) ){
+		std::cerr << "ERROR in LauFitObject::MultiDimConstraint : Different number of elements in vectors/covariance matrix!" << std::endl; 
+		gSystem->Exit( EXIT_FAILURE );
+	}
+
+	// Check invertion of the covariance matrix was successful
+	TMatrixD invCovMat {TMatrixD::kInverted, covMat};
+	if ( invCovMat == covMat ){
+		std::cerr << "ERROR in LauFitObject::MultiDimConstraint : covariance matrix inversion failed, check your input!" << std::endl; 
+		gSystem->Exit( EXIT_FAILURE );
+	}
+	invCovMat_ = invCovMat;
+
+	// Check invertion of the covariance matrix was successful
+	TDecompChol cholDecomp {covMat};
+	if ( ! cholDecomp.Decompose() ) {
+		std::cerr << "ERROR in LauFitObject::MultiDimConstraint : covariance matrix decomposition failed, check your input!" << std::endl; 
+		gSystem->Exit( EXIT_FAILURE );
+	}
+	sqrtCovMat_ = TMatrixD{TMatrixD::kTransposed, cholDecomp.GetU()};
+}
+
+Double_t LauFitObject::MultiDimConstraint::constraintPenalty() const
+{
+	TVectorD diff{ means_.GetNrows() };
+	for ( ULong_t j {0}; j < conLauPars_.size(); ++j ) {
+		LauParameter* param = conLauPars_[j];
+		diff[j] = param->unblindValue();
+	}
+	diff -= means_;
+	return 0.5 * invCovMat_.Similarity( diff );
+}
+
+void LauFitObject::MultiDimConstraint::generateConstraintMeans()
+{
+	TRandom* random = LauRandom::randomFun();
+
+	for ( Int_t j {0}; j < trueMeans_.GetNrows(); ++j ) {
+		means_[j] = random->Gaus(0.0, 1.0); 
+	}
+
+	means_ *= sqrtCovMat_;
+	means_ += trueMeans_;
 }